// lisa-frontend/src/api/types.ts
export interface Field {
  id: string
  name: string
  area: string | number
  location: string
  availableArea: string | number
  description?: string | null
  createdAt?: string
  updatedAt?: string
}

export interface Crop {
  id: string
  name: string
  type: string
  description?: string | null
  createdAt?: string
  updatedAt?: string
}
// ...
export interface HarvestStock {
  id: string
  cropId: string
  fieldId: string
  totalQuantity: string | number
  availableBalance: string | number
  unit: string
  crop?: Crop
  field?: Field
  createdAt?: string
  updatedAt?: string
}